javascript warm up

+++ 0x12-javascript-warm_up/9-add.js
@@ -2,10 +2,10 @@
 
 const { argv } = require('process');
 
-const first_phrase = parseFloat(argv[2]);
-const second_phrase = parseFloat(argv[3]);
+const first_phrase = parseInt(argv[2]);
+const second_phrase = parseInt(argv[3]);
 
 function add (a, b) {
   console.log(a + b);
 }
-  add(first_phrase, second_phrase);
+add(first_phrase, second_phrase);
